kbgitkbfs: Rename protocol to kbgitkbfs1.
