+++ libkbfs/data_types.go
@@ -15,7 +15,7 @@ import (
 	"github.com/keybase/kbfs/kbfsblock"
 	"github.com/keybase/kbfs/kbfscrypto"
 	"github.com/keybase/kbfs/kbfsmd"
-	kbgitkbfs "github.com/keybase/kbfs/protocol/kbgitkbfs"
+	kbgitkbfs "github.com/keybase/kbfs/protocol/kbgitkbfs1"
 	"github.com/keybase/kbfs/tlf"
 )
