+++ libkbfs/disk_block_cache_remote.go
@@ -12,7 +12,7 @@ import (
 	"github.com/keybase/go-framed-msgpack-rpc/rpc"
 	"github.com/keybase/kbfs/kbfsblock"
 	"github.com/keybase/kbfs/kbfscrypto"
-	kbgitkbfs "github.com/keybase/kbfs/protocol/kbgitkbfs"
+	kbgitkbfs "github.com/keybase/kbfs/protocol/kbgitkbfs1"
 	"github.com/keybase/kbfs/tlf"
 )
