+++ libkbfs/kbfs_service.go
@@ -10,7 +10,7 @@ import (
 	"github.com/keybase/client/go/logger"
 	keybase1 "github.com/keybase/client/go/protocol/keybase1"
 	"github.com/keybase/go-framed-msgpack-rpc/rpc"
-	kbgitkbfs "github.com/keybase/kbfs/protocol/kbgitkbfs"
+	kbgitkbfs "github.com/keybase/kbfs/protocol/kbgitkbfs1"
 )
 
 // KBFSErrorUnwrapper unwraps errors from the KBFS service.
